% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CerberusSim.R
\name{cerberus_sim}
\alias{cerberus_sim}
\title{Simulate from Cerberus Model}
\usage{
cerberus_sim(D1 = 5, D2 = 5, N = 10, Q = 2, true_priors = TRUE,
  use_names = TRUE)
}
\arguments{
\item{D1}{Number of multinomial categories in first dataset}

\item{D2}{Number of multinomial categories in second dataset}

\item{N}{number of samples}

\item{Q}{number of covariates}

\item{true_priors}{should Xi and upsilon be chosen to have mean at true
simulated value}

\item{use_names}{should outputs have named dimensions?}
}
\value{
list
}
\description{
Simulate from Cerberus Model
}
\examples{
sim <- cerberus_sim()
}
