% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{filter_sigma_ecdf}
\alias{filter_sigma_ecdf}
\title{Filter Sigma "Hits" Based on ECDF}
\usage{
filter_sigma_ecdf(Sigma, lambda = 0.95, use_names = TRUE)
}
\arguments{
\item{Sigma}{output of teacup_cerberus}

\item{lambda}{scalar between 0 and 1 (default = 0.95)}

\item{use_names}{if TRUE tries to name the hits rather than just giving indexes i,j}
}
\value{
data.frame of "hits" ranked by abs(ECDF(Sigma(i,j))(0)-0.5), first row is then
 the "hit" we are most confident is non-zero and large in magnitude.
}
\description{
Set a value of lambda  and this function will filter
to only show elements of Sigma(i,j) where i!=j that satisfy the following inequality:
ECDF(Sigma(i,j))(0) > lambda or  ECDF(Sigma(i,j))(0) < 1-lambda.
}
