% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{MultDirichletResample}
\alias{MultDirichletResample}
\title{Multinomial Dirichlet Sampler}
\usage{
MultDirichletResample(Y, alpha = rep(0.65, nrow(Y)), n_samples = 2000)
}
\arguments{
\item{Y}{count data (D x N) (e.g., taxa x samples)}

\item{alpha}{D-vector of Dirichlet Prior parameters}

\item{n_samples}{number of posterior samples to collect}
}
\value{
Posterior samples as array of dimensions: D x N x n_samples
}
\description{
Multinomial Dirichlet Sampler
}
