% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeacupCerberus-package.R
\docType{package}
\name{RcppCoDA}
\alias{RcppCoDA}
\alias{RcppCoDA-package}
\title{TeacupCerberus}
\description{
Estimates covariation within and between two count datasets where the counts
contain multinomial variation (e.g., sequence count data like microbiome 16S or bulk/single-cell RNA-seq).
The model outputs Bayesian posterior samples over covariance matricies.
The entire posterior reflects uncertainty in the true covariation due to multinomial counting.
}
\author{
Justin D Silverman
}
