% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{rDirichlet}
\alias{rDirichlet}
\title{Dirichlet Sampler}
\usage{
rDirichlet(alpha, n_samples = 2000)
}
\arguments{
\item{alpha}{D-vector of Dirichlet parameters}

\item{n_samples}{number of samples to produce}
}
\value{
array of samples of dimensions: D x n+samples
}
\description{
Dirichlet Sampler
}
