% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{tc2cor}
\alias{tc2cor}
\title{Convert Sigma from teacup_cerberus to Correlations}
\usage{
tc2cor(Sigma)
}
\arguments{
\item{Sigma}{Array of Covariance matricies from output of teacup_cerberus}
}
\value{
Array of Correlation matricies of same dimensions
}
\description{
Simple wrapper for cov2cor function that operates
on all posterior samples of Sigma
}
