% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{tc2iqlrvar}
\alias{tc2iqlrvar}
\title{Transform output of TeacupCerberus to iqlr}
\usage{
tc2iqlrvar(Sigma, D1, D2, qLow1 = 0.25, qHigh1 = 0.75, qLow2 = 0.25,
  qHigh2 = 0.75)
}
\arguments{
\item{Sigma}{output of TeacupCerberus (in CLR basis)}

\item{D1}{(integer) dimension of the first dataset}

\item{D2}{(integer) dimension of the second dataset}

\item{qLow1}{lower quantile defining iqlr for first dataset}

\item{qHigh1}{upper quantile defining iqlr for first dataset}

\item{qLow2}{lower quantile defining iqlr for second dataset}

\item{qHigh2}{upper quantile defining iqlr for second}
}
\value{
array of same dimensions as Sigma
}
\description{
Transform output of TeacupCerberus to iqlr
}
