% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{tc2phi}
\alias{tc2phi}
\title{Transform output of Teacup Cerberus to phi Statistics}
\usage{
tc2phi(Sigma)
}
\arguments{
\item{Sigma}{output of teacup_cerberus in (clr coordinates)}
}
\value{
array of the same dimensions as Sigma
}
\description{
Transform output of Teacup Cerberus to phi Statistics
}
\details{
this is just a wrapper around RcppCoDA::clrvar2phi
}
